split Verifier components
