+++ crates/kilt-dip-primitives/src/verifier/relaychain/mod.rs
@@ -20,7 +20,7 @@ use did::KeyIdOf;
 use frame_system::pallet_prelude::BlockNumberFor;
 use pallet_did_lookup::linkable_account::LinkableAccountId;
 use pallet_dip_consumer::{traits::IdentityProofVerifier, RuntimeCallOf};
-use pallet_dip_provider::{traits::IdentityCommitmentGenerator, IdentityCommitmentOf};
+use pallet_dip_provider::traits::IdentityCommitmentGenerator;
 use pallet_web3_names::Web3NameOf;
 use parity_scale_codec::{Decode, Encode};
 use scale_info::TypeInfo;
@@ -34,6 +34,8 @@ use crate::{
 	utils::OutputOf,
 	DipOriginInfo,
 };
+
+pub mod v0;
 
 mod error;
 pub use error::*;
@@ -177,225 +179,6 @@ impl<
 				identity_details,
 				v0_proof,
 			),
-		}
-	}
-}
-
-pub mod v0 {
-	use super::*;
-
-	use frame_support::ensure;
-	use frame_system::pallet_prelude::HeaderFor;
-	use sp_runtime::{traits::Zero, SaturatedConversion};
-
-	use crate::RelayDipDidProof;
-
-	/// Proof verifier configured given a specific KILT runtime implementation.
-	///
-	/// The generic types are the following:
-	///
-	/// * `ConsumerBlockHashStore`: A type providing block hashes for the
-	///   relaychain blocks.
-	/// * `KILT_PARA_ID`: The ID of the specific KILT parachain instance.
-	/// * `KiltRuntime`: A KILT runtime definition.
-	/// * `DidCallVerifier`: Logic to map `RuntimeCall`s to a specific DID key
-	///   relationship. This information is used once the Merkle proof is
-	///   verified, to filter only the revealed keys that match the provided
-	///   relationship.
-	/// * `SignedExtra`: Any additional information that must be signed by the
-	///   DID subject in the cross-chain operation.
-	/// * `MAX_PROVIDER_HEAD_PROOF_LEAVE_COUNT`: The maximum number of leaves
-	///   that can be revealed as part of the parachain head storage proof.
-	/// * `MAX_PROVIDER_HEAD_PROOF_LEAVE_SIZE`: The maximum size of each leaf
-	///   revealed as part of the parachain head storage proof.
-	/// * `MAX_DIP_COMMITMENT_PROOF_LEAVE_COUNT`: The maximum number of leaves
-	///   that can be revealed as part of the DIP commitment storage proof.
-	/// * `MAX_DIP_COMMITMENT_PROOF_LEAVE_SIZE`: The maximum size of each leaf
-	///   revealed as part of the DIP commitment storage proof.
-	/// * `MAX_DID_MERKLE_PROOF_LEAVE_COUNT`: The maximum number of *blinded*
-	///   leaves that can be revealed as part of the DID Merkle proof.
-	/// * `MAX_DID_MERKLE_PROOF_LEAVE_SIZE`: The maximum size of each *blinded*
-	///   leaf revealed as part of the DID Merkle proof.
-	/// * `MAX_DID_MERKLE_PROOF_LEAVE_SIZE`: The maximum number of leaves that
-	///   can be revealed as part of the DID Merkle proof.
-	pub struct RelaychainVerifier<
-		ConsumerBlockHashStore,
-		const KILT_PARA_ID: u32,
-		KiltRuntime,
-		DidCallVerifier,
-		SignedExtra,
-		const MAX_PROVIDER_HEAD_PROOF_LEAVE_COUNT: u32,
-		const MAX_PROVIDER_HEAD_PROOF_LEAVE_SIZE: u32,
-		const MAX_DIP_COMMITMENT_PROOF_LEAVE_COUNT: u32,
-		const MAX_DIP_COMMITMENT_PROOF_LEAVE_SIZE: u32,
-		const MAX_DID_MERKLE_PROOF_LEAVE_COUNT: u32,
-		const MAX_DID_MERKLE_PROOF_LEAVE_SIZE: u32,
-		const MAX_DID_MERKLE_LEAVES_REVEALED: u32,
-	>(
-		#[allow(clippy::type_complexity)]
-		PhantomData<(ConsumerBlockHashStore, KiltRuntime, DidCallVerifier, SignedExtra)>,
-	);
-
-	impl<
-			ConsumerRuntime,
-			ConsumerBlockHashStore,
-			const KILT_PARA_ID: u32,
-			KiltRuntime,
-			DidCallVerifier,
-			SignedExtra,
-			const MAX_PROVIDER_HEAD_PROOF_LEAVE_COUNT: u32,
-			const MAX_PROVIDER_HEAD_PROOF_LEAVE_SIZE: u32,
-			const MAX_DIP_COMMITMENT_PROOF_LEAVE_COUNT: u32,
-			const MAX_DIP_COMMITMENT_PROOF_LEAVE_SIZE: u32,
-			const MAX_DID_MERKLE_PROOF_LEAVE_COUNT: u32,
-			const MAX_DID_MERKLE_PROOF_LEAVE_SIZE: u32,
-			const MAX_DID_MERKLE_LEAVES_REVEALED: u32,
-		> IdentityProofVerifier<ConsumerRuntime>
-		for RelaychainVerifier<
-			ConsumerBlockHashStore,
-			KILT_PARA_ID,
-			KiltRuntime,
-			DidCallVerifier,
-			SignedExtra,
-			MAX_PROVIDER_HEAD_PROOF_LEAVE_COUNT,
-			MAX_PROVIDER_HEAD_PROOF_LEAVE_SIZE,
-			MAX_DIP_COMMITMENT_PROOF_LEAVE_COUNT,
-			MAX_DIP_COMMITMENT_PROOF_LEAVE_SIZE,
-			MAX_DID_MERKLE_PROOF_LEAVE_COUNT,
-			MAX_DID_MERKLE_PROOF_LEAVE_SIZE,
-			MAX_DID_MERKLE_LEAVES_REVEALED,
-		> where
-		ConsumerRuntime: pallet_dip_consumer::Config<Identifier = KiltRuntime::Identifier>,
-		ConsumerRuntime::LocalIdentityInfo: Incrementable + Default,
-		BlockNumberFor<ConsumerRuntime>: Into<U256> + TryFrom<U256>,
-		ConsumerBlockHashStore:
-			GetWithArg<BlockNumberFor<ConsumerRuntime>, Result = Option<OutputOf<ConsumerRuntime::Hashing>>>,
-		KiltRuntime: frame_system::Config<Hash = ConsumerRuntime::Hash>
-			+ pallet_dip_provider::Config
-			+ did::Config
-			+ pallet_web3_names::Config
-			+ pallet_did_lookup::Config,
-		KiltRuntime::IdentityCommitmentGenerator:
-			IdentityCommitmentGenerator<KiltRuntime, Output = ConsumerRuntime::Hash>,
-		IdentityCommitmentOf<KiltRuntime>: Into<KiltRuntime::Hash>,
-		SignedExtra: GetWithoutArg,
-		SignedExtra::Result: Encode,
-		DidCallVerifier: DipCallOriginFilter<
-			RuntimeCallOf<ConsumerRuntime>,
-			OriginInfo = RevealedDidKey<KeyIdOf<KiltRuntime>, BlockNumberFor<KiltRuntime>, KiltRuntime::AccountId>,
-		>,
-		DidCallVerifier::Error: Into<u8>,
-	{
-		type Error = DipRelaychainStateProofVerifierError<DidCallVerifier::Error>;
-		type Proof = RelayDipDidProof<
-			BlockNumberFor<ConsumerRuntime>,
-			ConsumerRuntime::Hashing,
-			KeyIdOf<KiltRuntime>,
-			KiltRuntime::AccountId,
-			BlockNumberFor<KiltRuntime>,
-			Web3NameOf<KiltRuntime>,
-			LinkableAccountId,
-		>;
-		type VerificationResult = DipOriginInfo<
-			KeyIdOf<KiltRuntime>,
-			KiltRuntime::AccountId,
-			BlockNumberFor<KiltRuntime>,
-			Web3NameOf<KiltRuntime>,
-			LinkableAccountId,
-			MAX_DID_MERKLE_LEAVES_REVEALED,
-		>;
-
-		fn verify_proof_for_call_against_details(
-			call: &RuntimeCallOf<ConsumerRuntime>,
-			subject: &ConsumerRuntime::Identifier,
-			submitter: &ConsumerRuntime::AccountId,
-			identity_details: &mut Option<ConsumerRuntime::LocalIdentityInfo>,
-			proof: Self::Proof,
-		) -> Result<Self::VerificationResult, Self::Error> {
-			// 1. Verify provided relaychain header.
-			let proof_without_header = proof
-				.verify_relay_header::<ConsumerBlockHashStore>()
-				.map_err(DipRelaychainStateProofVerifierError::ProofVerification)?;
-
-			// 2. Verify parachain state is finalized by relay chain and fresh.
-			ensure!(
-				proof_without_header.provider_head_proof.proof.len()
-					<= MAX_PROVIDER_HEAD_PROOF_LEAVE_COUNT.saturated_into(),
-				DipRelaychainStateProofVerifierError::ProofComponentTooLarge(0)
-			);
-			ensure!(
-				proof_without_header
-					.provider_head_proof
-					.proof
-					.iter()
-					.all(|l| l.len() <= MAX_PROVIDER_HEAD_PROOF_LEAVE_SIZE.saturated_into()),
-				DipRelaychainStateProofVerifierError::ProofComponentTooLarge(1)
-			);
-			let proof_without_relaychain = proof_without_header
-				.verify_provider_head_proof::<ConsumerRuntime::Hashing, HeaderFor<KiltRuntime>>(KILT_PARA_ID)
-				.map_err(DipRelaychainStateProofVerifierError::ProofVerification)?;
-
-			// 3. Verify commitment is included in provider parachain state.
-			ensure!(
-				proof_without_relaychain.dip_commitment_proof.0.len()
-					<= MAX_DIP_COMMITMENT_PROOF_LEAVE_COUNT.saturated_into(),
-				DipRelaychainStateProofVerifierError::ProofComponentTooLarge(2)
-			);
-			ensure!(
-				proof_without_relaychain
-					.dip_commitment_proof
-					.0
-					.iter()
-					.all(|l| l.len() <= MAX_DIP_COMMITMENT_PROOF_LEAVE_SIZE.saturated_into()),
-				DipRelaychainStateProofVerifierError::ProofComponentTooLarge(3)
-			);
-			let proof_without_parachain = proof_without_relaychain
-				.verify_dip_commitment_proof_for_subject::<KiltRuntime::Hashing, KiltRuntime>(subject)
-				.map_err(DipRelaychainStateProofVerifierError::ProofVerification)?;
-
-			// 4. Verify DIP Merkle proof.
-			ensure!(
-				proof_without_parachain.dip_proof.blinded.len() <= MAX_DID_MERKLE_PROOF_LEAVE_COUNT.saturated_into(),
-				DipRelaychainStateProofVerifierError::ProofComponentTooLarge(4)
-			);
-			ensure!(
-				proof_without_parachain
-					.dip_proof
-					.blinded
-					.iter()
-					.all(|l| l.len() <= MAX_DID_MERKLE_PROOF_LEAVE_SIZE.saturated_into()),
-				DipRelaychainStateProofVerifierError::ProofComponentTooLarge(5)
-			);
-			let proof_without_dip_merkle = proof_without_parachain
-				.verify_dip_proof::<KiltRuntime::Hashing, MAX_DID_MERKLE_LEAVES_REVEALED>()
-				.map_err(DipRelaychainStateProofVerifierError::ProofVerification)?;
-
-			// 5. Verify call is signed by one of the DID keys revealed in the proof
-			let current_block_number = frame_system::Pallet::<ConsumerRuntime>::block_number();
-			let consumer_genesis_hash =
-				frame_system::Pallet::<ConsumerRuntime>::block_hash(BlockNumberFor::<ConsumerRuntime>::zero());
-			let signed_extra = SignedExtra::get();
-			let encoded_payload = (call, &identity_details, submitter, consumer_genesis_hash, signed_extra).encode();
-			let revealed_did_info = proof_without_dip_merkle
-				.verify_signature_time(&current_block_number)
-				.and_then(|p| p.retrieve_signing_leaf_for_payload(&encoded_payload[..]))
-				.map_err(DipRelaychainStateProofVerifierError::ProofVerification)?;
-
-			// 6. Verify the signing key fulfills the requirements
-			let signing_key = revealed_did_info
-				.get_signing_leaf()
-				.map_err(DipRelaychainStateProofVerifierError::ProofVerification)?;
-			DidCallVerifier::check_call_origin_info(call, signing_key)
-				.map_err(DipRelaychainStateProofVerifierError::DidOriginError)?;
-
-			// 7. Increment the local details
-			if let Some(details) = identity_details {
-				details.increment();
-			} else {
-				*identity_details = Some(Default::default());
-			};
-
-			Ok(revealed_did_info)
 		}
 	}
 }
